pyproject.toml
[project]
name = "quantax"
version = "0.1.0"
description = "" 
readme = "README.md"
authors = [
    { name = "Yannik Mahlau", email = "mahlau@tnt.uni-hannover.de" },
]
requires-python = ">=3.11"
dependencies = [
    "pytreeclass>=0.11",
    "jax>=0.7.0",
    "seaborn>=0.13.2",
    "matplotlib>=3.9.4",
    "plum-dispatch>=2.5.7",
    "fastcore>=1.8.7",
    "jaxtyping>=0.3.7",
    "frozendict>=2.4.7",
    "rustworkx>=0.17.1",
]

[project.urls]
Repository = "https://github.com/ymahlau/quantax"
Documentation = "https://quantax.readthedocs.io/en/latest/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel.sources]
external_types = "typings"

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true
exclude = ["examples", "slurm", "checks", "docs"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "ipykernel>=6.29.5",
    "pre-commit>=4.0.1",
    "pymdown-extensions>=10.12",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.2",
    "tox-uv>=1.16.1",
    "pydoclint>=0.6.6",
    "sphinx>=8.0",
    "sphinx-autobuild>=2025.8.25",
    "sphinx-book-theme>=1.1.4",
    "nbsphinx>=0.9",
    "sphinx-autodoc-typehints>=3.2.0",
    "myst_nb>=1.3.0",
]

[tool.pyright]
include = ["src", "examples", "checks", "test"]
exclude = ["**/.venv"]
